requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
gradio==3.50.2
PyPDF2==3.0.1
pdf2image==1.16.3
pillow==10.2.0
sentence-transformers==2.2.2
chromadb==0.4.24
pytesseract==0.3.10
python-dotenv==1.0.1
